+++ scratch.js
@@ -11,16 +11,16 @@ spacetime.extend(require('./plugins/week-math/plugin.js'))
 //   console.log('   ' + wk.format('nice'))
 // })
 
-const byWeek = function (year) {
-  let arr = spacetime('jan 1 ' + year)
-    .minus(1, 'hour')
-    .every('week', 'Jan 1st 2022')
-  return arr.map((s) => {
-    let res = s.whichWeek()
-    return { title: s.format('nice-day'), num: res.num, month: res.month }
-  })
-}
-console.log(byWeek(2020))
+// const byWeek = function (year) {
+//   let arr = spacetime('jan 1 ' + year)
+//     .minus(1, 'hour')
+//     .every('week', 'Jan 1st 2022')
+//   return arr.map((s) => {
+//     let res = s.whichWeek()
+//     return { title: s.format('nice-day'), num: res.num, month: res.month }
+//   })
+// }
+// console.log(byWeek(2020))
 
 // const padEnd = function (str, width) {
 //   str = str.toString()
@@ -43,3 +43,13 @@ console.log(byWeek(2020))
 //     console.log(`\t${name}  - ${wk.format('{date-ordinal}')}-${sun.format('date-ordinal')}`)
 //   })
 // })
+
+// let d = spacetime('2020-11-01T01:32:00.000-06:00')
+// console.log(d.time()) // 1:32am
+// d = d.add(30, 'minutes')
+// console.log(d.time() === '2:02am') // 1:02am
+
+// let d = spacetime('december 1 2020').add(1, 'season')
+// let d = spacetime().endOf('season')
+let d = spacetime().next('season')
+console.log(d.format('nice'))