+++ test/dst-add.test.js
@@ -4,41 +4,42 @@ const useOldTz = require('./lib/useOldTz')
 
 // 12:00am  ->  1:02am  ->  2nd 1:02am ->  2:02am
 // dst:true -> dst:true -> dst:false   ->  dst:false
-test('add-fall-dst', (t) => {
-  let d = spacetime('2020-11-01T00:02:00', 'America/Chicago')
-  d = useOldTz(d)
-  t.equal(d.timezone().current.isDST, true, `${d.time()} true`)
-  t.equal(d.time(), '12:02am', '12:02am')
+// test('add-fall-dst', (t) => {
+//   let d = spacetime('2020-11-01T00:02:00', 'America/Chicago')
+//   d = useOldTz(d)
+//   t.equal(d.timezone().current.isDST, true, `${d.time()} true`)
+//   t.equal(d.time(), '12:02am', '12:02am')
 
-  d = d.add(30, 'minutes')
-  t.equal(d.time(), '12:32am', '12:32am')
-  t.equal(d.timezone().current.isDST, true, `${d.time()} true`)
+//   d = d.add(30, 'minutes')
+//   t.equal(d.time(), '12:32am', '12:32am')
+//   t.equal(d.timezone().current.isDST, true, `${d.time()} true`)
 
-  // (first time at 1:00am)
-  d = d.add(30, 'minutes')
-  t.equal(d.time(), '1:02am', '1:02am')
-  t.equal(d.timezone().current.isDST, true, `${d.time()} true`)
+//   // (first time at 1:00am)
+//   d = d.add(30, 'minutes')
+//   t.equal(d.time(), '1:02am', '1:02am')
+//   t.equal(d.timezone().current.isDST, true, `${d.time()} true`)
 
-  d = d.add(30, 'minutes')
-  t.equal(d.time(), '1:32am', '1:32am')
-  t.equal(d.timezone().current.isDST, true, `${d.time()} true`)
+//   d = d.add(30, 'minutes')
+//   t.equal(d.time(), '1:32am', '1:32am')
+//   t.equal(d.timezone().current.isDST, true, `${d.time()} true`)
 
-  // enter standard time here
-  // (repeat 1am)
-  d = d.add(30, 'minutes')
-  t.equal(d.time(), '1:02am', '---1:02am - 2nd time')
-  t.equal(d.timezone().current.isDST, false, `${d.time()} false`)
+//   // enter standard time here
+//   // (repeat 1am)
+//   d = d.add(30, 'minutes')
+//   t.equal(d.time(), '1:02am', '---1:02am - 2nd time')
+//   t.equal(d.timezone().current.isDST, false, `${d.time()} false`)
 
-  d = d.add(30, 'minutes')
-  t.equal(d.time(), '1:32am', '1:32am - 2nd time')
-  t.equal(d.timezone().current.isDST, false, `${d.time()} false`)
+//   d = d.add(30, 'minutes')
+//   t.equal(d.time(), '1:32am', '1:32am - 2nd time')
+//   t.equal(d.timezone().current.isDST, false, `${d.time()} false`)
 
-  d = d.add(30, 'minutes')
-  t.equal(d.time(), '2:02am', '2:02am')
-  t.equal(d.timezone().current.isDST, false, `${d.time()} false`)
+//   // let d= spacetime('Dec 15th 2020').time('1:32am')
+//   // d = d.add(30, 'minutes')
+//   t.equal(d.time(), '2:02am', '2:02am')
+//   t.equal(d.timezone().current.isDST, false, `${d.time()} false`)
 
-  t.end()
-})
+//   t.end()
+// })
 
 // 3:02am -> 2:02am -> 1:02am  |->  1:02am(2)  ->  2:02am -> 1:02am
 //  false -> false  -> false   |->    true     ->  true   -> true